requirements.txt
# System requirements:
# Ubuntu/Debian: sudo apt-get install portaudio19-dev python3-dev
# Fedora: sudo dnf install portaudio-devel
# macOS: brew install portaudio

streamlit
opencv-python
numpy
pytesseract
torch
transformers
pinecone-client
python-dotenv
langchain
langchain-community
openai
pandas
matplotlib
wordcloud
spacy
beautifulsoup4
requests
librosa
sounddevice
scipy
ctcdecode